requirements.txt
fastapi
pydantic
uvicorn
streamlit
requests
python-dotenv
pandas
datasets
kagglehub
langchain-core
langchain-chroma
langchain-huggingface
langchain-groq
langchain-text-splitters
sentence-transformers
pytest
